fix: reject fractional indices and step in std.slice (#797) (#884)
## Motivation

`std.slice` and the `[a:b:c]` slice operator silently truncated
fractional `index` / `end` / `step` arguments instead of rejecting them:

```jsonnet
std.slice([0, 1, 2], 0.9, 2.9, 1)   // was: [0, 1]
std.slice("abc", 0, 3, 1.5)         // was: "abc"
```

Official Jsonnet 0.22.0 rejects fractional array indices at
manifestation (`array index was not integer: 0.9`). Per the maintainer
decision in #797 ("go with the strict approach — if the function expects
an int, reject anything else"), this applies strict integer semantics to
**both** array and string slicing rather than truncating.

## Modification

- `Util.slice` now accepts raw `Option[Double]` and validates
integrality at a single enforcement point, erroring on `NaN` /
`Infinity` / non-whole values.
- The `std.slice` builtin (`SetModule`) and the `visitSlice` operator
(`Evaluator`) pass raw doubles through instead of pre-truncating to
`Int` (which hid the fractional part before validation could see it).
- Regression tests:
- `slice_strict_integer.jsonnet` — integers, negatives, `null`, and
whole doubles (`2.0`) still work, for both `std.slice` and the `[a:b:c]`
operator.
  - `error.slice_fractional_index.jsonnet` — fractional index rejected.

## Result

```
sjsonnet.Error: [std.slice] slice index must be an integer, got: 0.9
sjsonnet.Error: slice index must be an integer, got: 0.5    # operator path
```

Full JVM (Scala 3) test suite passes; formatting verified with
`checkFormat`.

Addresses #797.

+++ sjsonnet/src/sjsonnet/stdlib/SetModule.scala
@@ -471,7 +471,13 @@ object SetModule extends AbstractFunctionModule {
      * arr[0:4:1] instead of std.slice(arr, 0, 4, 1)).
      */
     builtin("slice", "indexable", "index", "end", "step") {
-      (pos, ev, indexable: Val, index: Option[Int], _end: Option[Int], _step: Option[Int]) =>
+      (
+          pos,
+          ev,
+          indexable: Val,
+          index: Option[Double],
+          _end: Option[Double],
+          _step: Option[Double]) =>
         Util.slice(pos, ev, indexable, index, _end, _step)
     },
     /**
